Move Context class into Flow module
There are several classes that seem to relate to the flow of filling out forms. This is the first file to be moved.
